COMMON-213: Generified Condition, Conflict, and Request with errors in creating Requests

+++ common-wpi-util/src/test/java/com/team2073/common/mediator/MediatorTestFixtures.java
@@ -27,7 +27,7 @@ public class MediatorTestFixtures {
 
         @Override
         public void set(Double place) {
-            currentPosition = (java.lang.Double) place;
+            currentPosition = place;
         }
 
         @Override
